+++ spec/db_helper.rb
@@ -6,6 +6,8 @@
 # This source code is licensed under the MIT license found in the
 # LICENSE file in the root directory of this source tree.
 #
+require 'mysql2'
+require_relative 'yaml_config_helpers'
 
 CLEAN_SQL_STATEMENTS = [
   'DELETE FROM `simmer_test`.`notes`',